+++ docs/src/neverworld.md
@@ -1,5 +1,5 @@
 # [List of functions in WenoNeverworld](@id sec:API)
 
 ```@autodocs
-Modules = [ WenoNeverworld, WenoNeverworld.Diagnostics]
+Modules = [ WenoNeverworld, WenoNeverworld.Diagnostics, WenoNeverworld.Auxiliaries, WenoNeverworld.NeverworldGrids, WenoNeverworld.NeverworldBoundaries, WenoNeverworld.Parameterizations]
 ```
