+++ src/Constants.jl
@@ -1,0 +1,13 @@
+module Constants
+
+####
+#### All the constants we need for the WenoNeverworld simulation
+####
+
+const Lz = 4000
+const Ly = 70
+const h  = 1000.0
+const ΔB = 6.0e-2 
+const max_latitude = 70
+
+end
